Write to csv only when wraparound is detected
