+++ zeus/device/cpu/rapl.py
@@ -13,21 +13,150 @@
 
 from __future__ import annotations
 
+import atexit
 import os
 import warnings
 from glob import glob
+import typing
 from typing import Sequence
 from functools import lru_cache
+import tempfile
+import multiprocessing as mp
+from time import time, sleep
+
+import pandas as pd
 
 import zeus.device.cpu.common as cpu_common
 from zeus.device.cpu.common import CpuDramMeasurement
 from zeus.device.exception import ZeusBaseCPUError
-from zeus.monitor.rapl import RaplMonitor
 from zeus.utils.logging import get_logger
 
 logger = get_logger(name=__name__)
 
 RAPL_DIR = "/sys/class/powercap/intel-rapl"
+
+
+class RaplWraparoundTracker:
+    """Monitor the wrapping around of RAPL counters.
+
+    This class acts as a lower level wrapper around a Python process that polls
+    the wrapping of RAPL counters. This is primarily used by
+    [`RAPLCPUs`][zeus.device.cpu.rapl.RAPLCPUs].
+
+    !!! Warning
+        Since the monitor spawns a child process, **it should not be instantiated as a global variable**.
+        Python puts a protection to prevent creating a process in global scope.
+        Refer to the "Safe importing of main module" section in the
+        [Python documentation](https://docs.python.org/3/library/multiprocessing.html#the-spawn-and-forkserver-start-methods)
+        for more details.
+
+    Attributes:
+        rapl_file_path (str): File path of rapl file to track wraparounds for.
+        max_energy_uj (float): Max value of rapl counter for `rapl_file_path` file. Used to
+        determine the sleep period between polls
+    """
+
+    def __init__(
+        self,
+        rapl_file_path: str,
+        max_energy_uj: float,
+        rapl_csv_path: str | None = None,
+    ) -> None:
+        """Initialize the rapl monitor.
+
+        Args:
+            rapl_file_path: File path where the RAPL file is located
+            max_energy_uj: Max energy range uj value
+            rapl_csv_path: If given, the wrap around polling will write measurements
+                to this path. Otherwise, a temporary file will be used.
+        """
+        if not os.path.exists(rapl_file_path):
+            raise ValueError(f"{rapl_file_path} is not a valid file path")
+        self.rapl_file_path = rapl_file_path
+
+        # Set up logging.
+        self.logger = get_logger(type(self).__name__)
+
+        self.logger.info("Monitoring wrap around of %s", rapl_file_path)
+
+        # Create and open the CSV to record power measurements.
+        if rapl_csv_path is None:
+            rapl_csv_path = tempfile.mkstemp(suffix=".csv", text=True)[1]
+        open(rapl_csv_path, "w").close()
+        self.rapl_f = open(rapl_csv_path)
+        self.rapl_df_columns = ["time", "energy"]
+        self.rapl_df = pd.DataFrame(columns=self.rapl_df_columns)
+
+        # Spawn the power polling process.
+        atexit.register(self._stop)
+        self.process = mp.get_context("spawn").Process(
+            target=_polling_process,
+            args=(rapl_file_path, max_energy_uj, rapl_csv_path),
+        )
+        self.process.start()
+
+    def _stop(self) -> None:
+        """Stop monitoring power usage."""
+        if self.process is not None:
+            self.process.terminate()
+            self.process.join(timeout=1.0)
+            self.process.kill()
+            self.process = None
+
+    def _update_df(self) -> None:
+        """Add rows to the power dataframe from the CSV file."""
+        try:
+            additional_df = typing.cast(
+                pd.DataFrame,
+                pd.read_csv(self.rapl_f, header=None, names=self.rapl_df_columns),
+            )
+        except pd.errors.EmptyDataError:
+            return
+
+        if additional_df.empty:
+            return
+
+        if self.rapl_df.empty:
+            self.rapl_df = additional_df
+        else:
+            self.rapl_df = pd.concat(
+                [self.rapl_df, additional_df],
+                axis=0,
+                ignore_index=True,
+                copy=False,
+            )
+
+    def get_num_wraparounds(self) -> int:
+        """Get the number of wraparounds detected by the polling process."""
+        self._update_df()
+        print(self.rapl_df)
+        return len(self.rapl_df)
+
+
+def _polling_process(
+    rapl_file_path: str,
+    max_energy_uj: float,
+    rapl_csv_path: str,
+) -> None:
+    """Run the rapl monitor."""
+    try:
+        # Use line buffering.
+        with open(rapl_file_path, "r") as rapl_file:
+            last_energy_uj = float(rapl_file.read().strip())
+        with open(rapl_csv_path, "w", buffering=1) as rapl_f:
+            while True:
+                now = time()
+                sleep_time = 1.0
+                with open(rapl_file_path, "r") as rapl_file:
+                    energy_uj = float(rapl_file.read().strip())
+                    if max_energy_uj - energy_uj < 1000:
+                        sleep_time = 0.1
+                if energy_uj < last_energy_uj:
+                    rapl_f.write(f"{now},{energy_uj}\n")
+                last_energy_uj = energy_uj
+                sleep(sleep_time)
+    except KeyboardInterrupt:
+        return
 
 
 @lru_cache(maxsize=1)
@@ -87,7 +216,9 @@ class RAPLFile:
                 "Error reading package max energy range"
             ) from err
 
-        self.monitor = RaplMonitor(self.energy_uj_path, self.max_energy_range_uj)
+        self.wraparound_tracker = RaplWraparoundTracker(
+            self.energy_uj_path, self.max_energy_range_uj
+        )
 
     def __str__(self) -> str:
         """Return a string representation of the RAPL file object."""
@@ -102,8 +233,8 @@ class RAPLFile:
         """
         with open(self.energy_uj_path) as energy_file:
             new_energy_uj = float(energy_file.read().strip())
-        num_wraparounds = self.monitor.get_num_wraparounds()
-        return (new_energy_uj + num_wraparounds*self.max_energy_range_uj) / 1000.0
+        num_wraparounds = self.wraparound_tracker.get_num_wraparounds()
+        return (new_energy_uj + num_wraparounds * self.max_energy_range_uj) / 1000.0
 
 
 class RAPLCPU(cpu_common.CPU):
